linted to new standards
