+++ anzeigen-frontend/src/components/Filter/SearchAd.vue
@@ -13,6 +13,7 @@
 </template>
 
 <script setup lang="ts">
+// eslint-disable-next-line @typescript-eslint/no-empty-function
 const search = () => {};
 </script>
