+++ anzeigen-frontend/src/types/SortingOrderCriteria.ts
@@ -1,13 +1,13 @@
 type Criteria = "titel" | "price" | "creationDate";
 
-type CriteriaValue = {
+interface CriteriaValue {
   criteria: Criteria;
   order: "asc" | "desc";
-};
+}
 
-type SortingOrder = {
+interface SortingOrder {
   title: string;
   value: CriteriaValue;
-};
+}
 
 export type { Criteria, CriteriaValue, SortingOrder };
