+++ anzeigen-frontend/src/types/api/AdCategory.ts
@@ -1,5 +1,5 @@
-export type AdCategory = {
+export interface AdCategory {
   id: number;
   name: string;
   standard: boolean;
-};
+}
